Fix error accessing subspaces callouts (#6058)
